👌 IMPROVE: Rename to case sensitive directories
